# api/swagger/swagger.yaml
basePath: /api/v1
definitions:
  dto.GetRequest:
    properties:
      id:
        type: integer
    type: object
  dto.SaveRequest:
    properties:
      id:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      serverTime:
        type: string
      status:
        type: string
      success:
        type: boolean
    type: object
host: id.greet.org
info:
  contact:
    email: '[Your Email]'
    name: '[Your Name]'
    url: https://www.swagger.io/support
  description: Greet service API specification.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: API Specification
  version: "1.0"
paths:
  /example:
    get:
      consumes:
      - application/json
      description: GetEntity example by ID.
      parameters:
      - description: GetEntity Request
        in: body
        name: GetRequest
        required: true
        schema:
          $ref: '#/definitions/dto.GetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: GetEntity example
      tags:
      - example
    post:
      consumes:
      - application/json
      description: SaveEntity example.
      parameters:
      - description: SaveEntity Request
        in: body
        name: SaveRequest
        required: true
        schema:
          $ref: '#/definitions/dto.SaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: SaveEntity example
      tags:
      - example
schemes:
- http
- https
swagger: "2.0"